requirements.txt
timm==0.5.4
opencv-python==4.4.0.46
termcolor==1.1.0
yacs==0.1.8
pyyaml
scipy
einops
fvcore